configure.ac: change regex for sed version
Change regex for sed version in configure.ac.

This regex will match only on the first line and doesn't
look for a 'version' pattern which is not available in
sed version 4.2.2.

The new regex will match in the first line of 'sed --version' in a pattern
which starts with random character sequence(or zero) followed by whitespace
character and several digits of pattern x.x....x which x is a number.
This pattern will be save in the SED_VERSION variable.

Signed-off-by: Alexander Aring <alex.aring@gmail.com>
[mkl: add match for GNU in regex]
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>

(cherry picked from commit 2a89985c97422af3569eb6d268e3ccdc9c3fb0e5)
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>