images_oxnasblob: support generating a disk image bootable for OXNAS systems
On OXNAS systems, we can boot from SATA which contains a prepared
OXNAS boot area in front of the first partition and two OXNAS boot
parameter entries inside the master boot record that will be used
from the on-chip OXNAS boot rom to load and start program code
from SATA.

The boot area is devided into a primary and a secondary part which
each one contains the bootloader U-Boot and the Linux kernel image.
The bootloader have to be packaged as an OXNAS payload blobs that
can be accepted by the OXNAS boot rom.

Signed-off-by: Stephan Linz <linz@li-pro.net>